feat(task-db): migrate task_db from SQLite to Postgres (closes #750)
Replace the SQLite-backed TaskDb with a Postgres implementation:

- Add harness_core::db_pg with pg_open_pool, pg_open_pool_schematized,
  and PgMigrator (transactional DDL, idempotent schema_migrations tracking)
- Rewrite task_db as a module with typed sub-files (migrations, queries_tasks,
  queries_aux, types); bind syntax ?->$N, dialect conversions throughout
- Schema isolation via h{hash(db_path)} namespace — each unique db_path gets
  its own Postgres schema, giving integration tests full parallel isolation
  without cfg(test) gating (cfg(test) is not set for library deps of
  integration-test binaries)
- Fix INTEGER->BIGINT in v1/v5/v12/v17 migrations to match Rust i64 fields
- Add v18 version column for optimistic locking
- Update periodic_retry.rs test SQL from SQLite datetime() to Postgres NOW()
- Update pre-commit hook to skip Postgres-dependent lib tests when
  DATABASE_URL is not set; CI always has DATABASE_URL configured
